Added a wins/losses counter
Added a counter for wins and losses for whenever the user restarts the game.

Also fixed some issues with the messages, for example, if a person input a non-existing word they received a message that the word does not exist and the message still remained after they inputted a valid word. This is no longer the case.